Refactor history database migrations
