+++ crates/agent-gui/src-tauri/src/commands/chat_history.rs
@@ -4,15 +4,13 @@ use rusqlite::{params, Connection, OptionalExtension};
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 use std::{
-    collections::{HashMap, HashSet},
-    fs,
-    path::PathBuf,
+    collections::HashMap,
     sync::Arc,
     time::{Duration, SystemTime, UNIX_EPOCH},
 };
 
 use crate::{
-    commands::subagent_history,
+    commands::{history_db, subagent_history},
     services::{
         gateway::{build_history_sync_delete, build_history_sync_upsert, GatewayController},
         memory::{MemoryHistorySearchMatch, MemorySearchArgs},
@@ -20,7 +18,6 @@ use crate::{
 };
 use uuid::Uuid;
 
-const DB_FILENAME: &str = "chat-history.sqlite3";
 const HISTORY_SHARE_TOKEN_LEN: usize = 9;
 const HISTORY_SHARE_TOKEN_INSERT_ATTEMPTS: usize = 8;
 const HISTORY_SHARE_TOKEN_ALPHABET: &[u8] =
@@ -245,461 +242,8 @@ fn now_ms() -> i64 {
     duration.as_millis() as i64
 }
 
-fn chat_history_dir() -> Result<PathBuf, String> {
-    let home = dirs::home_dir().ok_or_else(|| "无法定位用户目录".to_string())?;
-    let dir = home.join(format!(".{}", env!("CARGO_PKG_NAME")));
-    fs::create_dir_all(&dir).map_err(|e| format!("创建历史目录失败：{e}"))?;
-    Ok(dir)
-}
-
 fn open_db() -> Result<Connection, String> {
-    let db_path = chat_history_dir()?.join(DB_FILENAME);
-    let conn = Connection::open(db_path).map_err(|e| format!("打开聊天历史数据库失败：{e}"))?;
-    conn.busy_timeout(Duration::from_secs(5))
-        .map_err(|e| format!("设置 SQLite busy_timeout 失败：{e}"))?;
-    initialize_db(&conn)?;
-
-    Ok(conn)
-}
-
-fn initialize_db(conn: &Connection) -> Result<(), String> {
-    conn.execute_batch(
-        "
-        PRAGMA foreign_keys = ON;
-
-        CREATE TABLE IF NOT EXISTS chatHistory (
-            id TEXT PRIMARY KEY,
-            title TEXT NOT NULL,
-            provider_id TEXT NOT NULL,
-            model TEXT NOT NULL,
-            session_id TEXT,
-            cwd TEXT,
-            context_meta_json TEXT,
-            active_segment_index INTEGER,
-            total_segment_count INTEGER,
-            total_message_count INTEGER,
-            created_at INTEGER NOT NULL,
-            updated_at INTEGER NOT NULL,
-            is_pinned INTEGER NOT NULL DEFAULT 0,
-            pinned_at INTEGER
-        );
-
-        CREATE TABLE IF NOT EXISTS chatHistorySegment (
-            conversation_id TEXT NOT NULL,
-            segment_index INTEGER NOT NULL,
-            segment_id TEXT NOT NULL,
-            summary_json TEXT,
-            messages_json TEXT NOT NULL,
-            message_count INTEGER NOT NULL,
-            start_message_id TEXT,
-            end_message_id TEXT,
-            created_at INTEGER NOT NULL,
-            updated_at INTEGER NOT NULL,
-            PRIMARY KEY (conversation_id, segment_index),
-            UNIQUE (conversation_id, segment_id),
-            FOREIGN KEY (conversation_id) REFERENCES chatHistory(id) ON DELETE CASCADE
-        );
-
-        CREATE TABLE IF NOT EXISTS chatHistoryShare (
-            conversation_id TEXT PRIMARY KEY,
-            token TEXT UNIQUE,
-            enabled INTEGER NOT NULL DEFAULT 0,
-            redact_tool_content INTEGER NOT NULL DEFAULT 0,
-            created_at INTEGER NOT NULL,
-            updated_at INTEGER NOT NULL,
-            FOREIGN KEY (conversation_id) REFERENCES chatHistory(id) ON DELETE CASCADE
-        );
-        ",
-    )
-    .map_err(|e| format!("初始化聊天历史表失败：{e}"))?;
-
-    ensure_chat_history_columns(conn)?;
-    ensure_chat_history_segment_columns(conn)?;
-    ensure_chat_history_share_columns(conn)?;
-    conn.execute_batch(
-        "
-        CREATE INDEX IF NOT EXISTS idx_chatHistory_updated_at
-            ON chatHistory(updated_at DESC);
-        CREATE INDEX IF NOT EXISTS idx_chatHistorySegment_conversation_updated
-            ON chatHistorySegment(conversation_id, updated_at DESC);
-        CREATE INDEX IF NOT EXISTS idx_chatHistory_pinned
-            ON chatHistory(is_pinned DESC, pinned_at DESC, updated_at DESC);
-        CREATE INDEX IF NOT EXISTS idx_chatHistoryShare_token
-            ON chatHistoryShare(token);
-        ",
-    )
-    .map_err(|e| format!("初始化聊天历史置顶索引失败：{e}"))?;
-    ensure_chat_history_fts(conn)?;
-
-    Ok(())
-}
-
-fn read_table_columns(
-    conn: &Connection,
-    table_name: &str,
-    table_label: &str,
-) -> Result<HashSet<String>, String> {
-    let mut stmt = conn
-        .prepare(&format!("PRAGMA table_info({table_name})"))
-        .map_err(|e| format!("读取{table_label}结构失败：{e}"))?;
-    let rows = stmt
-        .query_map([], |row| row.get::<_, String>(1))
-        .map_err(|e| format!("查询{table_label}结构失败：{e}"))?;
-    let mut columns = HashSet::new();
-    for row in rows {
-        columns.insert(row.map_err(|e| format!("读取{table_label}字段失败：{e}"))?);
-    }
-    Ok(columns)
-}
-
-fn ensure_table_columns(
-    conn: &Connection,
-    table_name: &str,
-    table_label: &str,
-    migrations: &[(&str, &str)],
-) -> Result<(), String> {
-    let columns = read_table_columns(conn, table_name, table_label)?;
-
-    for (column, ddl) in migrations {
-        if !columns.contains(*column) {
-            conn.execute_batch(ddl)
-                .map_err(|e| format!("迁移{table_label}字段 {column} 失败：{e}"))?;
-        }
-    }
-
-    Ok(())
-}
-
-fn ensure_chat_history_columns(conn: &Connection) -> Result<(), String> {
-    ensure_table_columns(
-        conn,
-        "chatHistory",
-        "聊天历史主表",
-        &[
-            (
-                "title",
-                "ALTER TABLE chatHistory ADD COLUMN title TEXT NOT NULL DEFAULT 'Untitled';",
-            ),
-            (
-                "provider_id",
-                "ALTER TABLE chatHistory ADD COLUMN provider_id TEXT NOT NULL DEFAULT '';",
-            ),
-            (
-                "model",
-                "ALTER TABLE chatHistory ADD COLUMN model TEXT NOT NULL DEFAULT '';",
-            ),
-            (
-                "session_id",
-                "ALTER TABLE chatHistory ADD COLUMN session_id TEXT;",
-            ),
-            ("cwd", "ALTER TABLE chatHistory ADD COLUMN cwd TEXT;"),
-            (
-                "context_meta_json",
-                "ALTER TABLE chatHistory ADD COLUMN context_meta_json TEXT NOT NULL DEFAULT '{}';",
-            ),
-            (
-                "active_segment_index",
-                "ALTER TABLE chatHistory ADD COLUMN active_segment_index INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "total_segment_count",
-                "ALTER TABLE chatHistory ADD COLUMN total_segment_count INTEGER NOT NULL DEFAULT 1;",
-            ),
-            (
-                "total_message_count",
-                "ALTER TABLE chatHistory ADD COLUMN total_message_count INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "created_at",
-                "ALTER TABLE chatHistory ADD COLUMN created_at INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "updated_at",
-                "ALTER TABLE chatHistory ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "is_pinned",
-                "ALTER TABLE chatHistory ADD COLUMN is_pinned INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "pinned_at",
-                "ALTER TABLE chatHistory ADD COLUMN pinned_at INTEGER;",
-            ),
-        ],
-    )?;
-
-    conn.execute_batch(
-        "
-        UPDATE chatHistory
-        SET title = 'Untitled'
-        WHERE title IS NULL OR trim(title) = '';
-
-        UPDATE chatHistory
-        SET provider_id = ''
-        WHERE provider_id IS NULL;
-
-        UPDATE chatHistory
-        SET model = ''
-        WHERE model IS NULL;
-
-        UPDATE chatHistory
-        SET context_meta_json = '{}'
-        WHERE context_meta_json IS NULL OR trim(context_meta_json) = '';
-
-        UPDATE chatHistory
-        SET active_segment_index = 0
-        WHERE active_segment_index IS NULL OR active_segment_index < 0;
-
-        UPDATE chatHistory
-        SET total_segment_count = 1
-        WHERE total_segment_count IS NULL OR total_segment_count < 1;
-
-        UPDATE chatHistory
-        SET total_message_count = 0
-        WHERE total_message_count IS NULL OR total_message_count < 0;
-
-        UPDATE chatHistory
-        SET created_at = 0
-        WHERE created_at IS NULL;
-
-        UPDATE chatHistory
-        SET updated_at = created_at
-        WHERE updated_at IS NULL;
-
-        UPDATE chatHistory
-        SET is_pinned = 0
-        WHERE is_pinned IS NULL;
-        ",
-    )
-    .map_err(|e| format!("修复聊天历史主表默认字段失败：{e}"))?;
-
-    Ok(())
-}
-
-fn ensure_chat_history_segment_columns(conn: &Connection) -> Result<(), String> {
-    ensure_table_columns(
-        conn,
-        "chatHistorySegment",
-        "聊天历史分段表",
-        &[
-            (
-                "segment_id",
-                "ALTER TABLE chatHistorySegment ADD COLUMN segment_id TEXT NOT NULL DEFAULT '';",
-            ),
-            (
-                "summary_json",
-                "ALTER TABLE chatHistorySegment ADD COLUMN summary_json TEXT;",
-            ),
-            (
-                "messages_json",
-                "ALTER TABLE chatHistorySegment ADD COLUMN messages_json TEXT NOT NULL DEFAULT '[]';",
-            ),
-            (
-                "message_count",
-                "ALTER TABLE chatHistorySegment ADD COLUMN message_count INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "start_message_id",
-                "ALTER TABLE chatHistorySegment ADD COLUMN start_message_id TEXT;",
-            ),
-            (
-                "end_message_id",
-                "ALTER TABLE chatHistorySegment ADD COLUMN end_message_id TEXT;",
-            ),
-            (
-                "created_at",
-                "ALTER TABLE chatHistorySegment ADD COLUMN created_at INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "updated_at",
-                "ALTER TABLE chatHistorySegment ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0;",
-            ),
-        ],
-    )?;
-
-    conn.execute_batch(
-        "
-        UPDATE chatHistorySegment
-        SET segment_id = 'segment-' || segment_index
-        WHERE segment_id IS NULL OR trim(segment_id) = '';
-
-        UPDATE chatHistorySegment
-        SET messages_json = '[]'
-        WHERE messages_json IS NULL OR trim(messages_json) = '';
-
-        UPDATE chatHistorySegment
-        SET message_count = 0
-        WHERE message_count IS NULL OR message_count < 0;
-
-        UPDATE chatHistorySegment
-        SET created_at = 0
-        WHERE created_at IS NULL;
-
-        UPDATE chatHistorySegment
-        SET updated_at = created_at
-        WHERE updated_at IS NULL;
-        ",
-    )
-    .map_err(|e| format!("修复聊天历史分段表默认字段失败：{e}"))?;
-
-    Ok(())
-}
-
-fn ensure_chat_history_share_columns(conn: &Connection) -> Result<(), String> {
-    ensure_table_columns(
-        conn,
-        "chatHistoryShare",
-        "聊天历史分享表",
-        &[
-            ("token", "ALTER TABLE chatHistoryShare ADD COLUMN token TEXT;"),
-            (
-                "enabled",
-                "ALTER TABLE chatHistoryShare ADD COLUMN enabled INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "redact_tool_content",
-                "ALTER TABLE chatHistoryShare ADD COLUMN redact_tool_content INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "created_at",
-                "ALTER TABLE chatHistoryShare ADD COLUMN created_at INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "updated_at",
-                "ALTER TABLE chatHistoryShare ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0;",
-            ),
-        ],
-    )?;
-
-    conn.execute_batch(
-        "
-        UPDATE chatHistoryShare
-        SET enabled = 0
-        WHERE enabled IS NULL;
-
-        UPDATE chatHistoryShare
-        SET redact_tool_content = 0
-        WHERE redact_tool_content IS NULL;
-
-        UPDATE chatHistoryShare
-        SET created_at = 0
-        WHERE created_at IS NULL;
-
-        UPDATE chatHistoryShare
-        SET updated_at = created_at
-        WHERE updated_at IS NULL;
-        ",
-    )
-    .map_err(|e| format!("修复聊天历史分享表默认字段失败：{e}"))?;
-
-    Ok(())
-}
-
-fn ensure_chat_history_fts(conn: &Connection) -> Result<(), String> {
-    conn.execute_batch(
-        r#"
-        CREATE VIRTUAL TABLE IF NOT EXISTS chatHistorySegmentFts USING fts5(
-            conversation_id         UNINDEXED,
-            segment_index           UNINDEXED,
-            segment_id              UNINDEXED,
-            title,
-            cwd,
-            body,
-            segment_updated_at      UNINDEXED,
-            conversation_updated_at UNINDEXED,
-            tokenize = "trigram"
-        );
-
-        CREATE VIRTUAL TABLE IF NOT EXISTS chatHistoryMessageFts USING fts5(
-            conversation_id         UNINDEXED,
-            segment_index           UNINDEXED,
-            segment_id              UNINDEXED,
-            message_index           UNINDEXED,
-            message_id              UNINDEXED,
-            role                    UNINDEXED,
-            title,
-            cwd,
-            body,
-            message_updated_at      UNINDEXED,
-            segment_updated_at      UNINDEXED,
-            conversation_updated_at UNINDEXED,
-            tokenize = "trigram"
-        );
-
-        CREATE TABLE IF NOT EXISTS chatHistoryFtsSegmentIndex (
-            conversation_id         TEXT NOT NULL,
-            segment_index           INTEGER NOT NULL,
-            segment_updated_at      INTEGER NOT NULL,
-            conversation_updated_at INTEGER NOT NULL,
-            PRIMARY KEY (conversation_id, segment_index)
-        );
-        "#,
-    )
-    .map_err(|e| format!("初始化聊天历史 FTS 表失败：{e}"))?;
-
-    ensure_chat_history_fts_index_columns(conn)?;
-    conn.execute_batch(
-        "
-        CREATE INDEX IF NOT EXISTS idx_chatHistoryFtsSegmentIndex_segment_updated
-            ON chatHistoryFtsSegmentIndex(segment_updated_at DESC);
-
-        CREATE INDEX IF NOT EXISTS idx_chatHistoryFtsSegmentIndex_conversation_updated
-            ON chatHistoryFtsSegmentIndex(conversation_updated_at DESC);
-        ",
-    )
-    .map_err(|e| format!("初始化聊天历史 FTS 索引失败：{e}"))?;
-
-    seed_existing_chat_history_fts_index(conn)?;
-
-    Ok(())
-}
-
-fn ensure_chat_history_fts_index_columns(conn: &Connection) -> Result<(), String> {
-    ensure_table_columns(
-        conn,
-        "chatHistoryFtsSegmentIndex",
-        "聊天历史 FTS 元数据表",
-        &[
-            (
-                "segment_updated_at",
-                "ALTER TABLE chatHistoryFtsSegmentIndex ADD COLUMN segment_updated_at INTEGER NOT NULL DEFAULT 0;",
-            ),
-            (
-                "conversation_updated_at",
-                "ALTER TABLE chatHistoryFtsSegmentIndex ADD COLUMN conversation_updated_at INTEGER NOT NULL DEFAULT 0;",
-            ),
-        ],
-    )
-}
-
-fn seed_existing_chat_history_fts_index(conn: &Connection) -> Result<(), String> {
-    conn.execute(
-        "
-        INSERT OR IGNORE INTO chatHistoryFtsSegmentIndex (
-            conversation_id,
-            segment_index,
-            segment_updated_at,
-            conversation_updated_at
-        )
-        SELECT
-            f.conversation_id,
-            CAST(f.segment_index AS INTEGER),
-            s.updated_at,
-            h.updated_at
-        FROM chatHistorySegmentFts f
-        JOIN chatHistorySegment s
-          ON s.conversation_id = f.conversation_id
-         AND s.segment_index = CAST(f.segment_index AS INTEGER)
-        JOIN chatHistory h ON h.id = s.conversation_id
-        WHERE CAST(f.segment_updated_at AS INTEGER) = s.updated_at
-          AND CAST(f.conversation_updated_at AS INTEGER) = h.updated_at
-        GROUP BY f.conversation_id, CAST(f.segment_index AS INTEGER)
-        ",
-        [],
-    )
-    .map_err(|e| format!("同步历史 FTS 元数据失败：{e}"))?;
-
-    Ok(())
+    history_db::open_connection()
 }
 
 fn refresh_chat_history_fts(conn: &Connection, filter: &HistorySearchFilter) -> Result<(), String> {
@@ -3638,7 +3182,7 @@ mod tests {
             Connection::open_in_memory().map_err(|e| format!("打开测试聊天历史数据库失败：{e}"))?;
         conn.busy_timeout(Duration::from_secs(5))
             .map_err(|e| format!("设置测试 SQLite busy_timeout 失败：{e}"))?;
-        initialize_db(&conn)?;
+        history_db::initialize_connection(&conn)?;
         Ok(conn)
     }
 
@@ -3766,7 +3310,7 @@ mod tests {
         )
         .expect("create legacy chatHistory table");
 
-        initialize_db(&conn).expect("migrate legacy schema");
+        history_db::initialize_connection(&conn).expect("migrate legacy schema");
 
         let is_pinned_exists: i64 = conn
             .query_row(
@@ -3829,7 +3373,7 @@ mod tests {
         )
         .expect("create legacy chatHistory table");
 
-        initialize_db(&conn).expect("migrate legacy schema");
+        history_db::initialize_connection(&conn).expect("migrate legacy schema");
 
         let summaries = list_chat_history_sync(&conn, 1, 20).expect("list migrated legacy history");
         assert_eq!(summaries.total_count, 1);
@@ -3846,6 +3390,54 @@ mod tests {
         assert_eq!(record.active_segment_index, 0);
         assert_eq!(record.total_segment_count, 1);
         assert_eq!(record.total_message_count, 0);
+    }
+
+    #[test]
+    fn initialize_db_tolerates_case_variant_existing_columns() {
+        let conn =
+            Connection::open_in_memory().expect("open legacy in-memory chat history database");
+        conn.execute_batch(
+            "
+            CREATE TABLE chatHistory (
+                id TEXT PRIMARY KEY,
+                title TEXT NOT NULL,
+                provider_id TEXT NOT NULL,
+                model TEXT NOT NULL,
+                Context_Meta_Json TEXT,
+                created_at INTEGER NOT NULL,
+                updated_at INTEGER NOT NULL
+            );
+
+            INSERT INTO chatHistory (
+                id,
+                title,
+                provider_id,
+                model,
+                created_at,
+                updated_at
+            ) VALUES (
+                'legacy-conv',
+                'Legacy Conversation',
+                'codex',
+                'gpt-5',
+                1700000000000,
+                1700000000100
+            );
+            ",
+        )
+        .expect("create legacy chatHistory table with case-variant context meta column");
+
+        history_db::initialize_connection(&conn)
+            .expect("migrate legacy schema with case-variant column");
+
+        let context_meta_json: String = conn
+            .query_row(
+                "SELECT context_meta_json FROM chatHistory WHERE id = 'legacy-conv'",
+                [],
+                |row| row.get(0),
+            )
+            .expect("query migrated context meta");
+        assert_eq!(context_meta_json, "{}");
     }
 
     #[test]
@@ -3889,7 +3481,7 @@ mod tests {
             )
             .expect("create minimal legacy history schema");
 
-        initialize_db(&legacy).expect("migrate minimal legacy schema");
+        history_db::initialize_connection(&legacy).expect("migrate minimal legacy schema");
 
         for table_name in [
             "chatHistory",
@@ -4476,7 +4068,7 @@ mod tests {
         )
         .expect("insert legacy segment without fts");
 
-        initialize_db(&conn).expect("re-run schema initialization");
+        history_db::initialize_connection(&conn).expect("re-run schema initialization");
         let after_init =
             search_chat_history_fts(&conn, "热路径", 8, &default_history_search_filter())
                 .expect("search after schema init");
@@ -4699,7 +4291,7 @@ mod tests {
     }
 
     #[test]
-    fn subagent_prune_initializes_schema_on_chat_history_connection() {
+    fn subagent_prune_uses_initialized_history_schema() {
         let conn = open_test_db().expect("open test db");
         let before: Option<String> = conn
             .query_row(
@@ -4709,11 +4301,11 @@ mod tests {
             )
             .optional()
             .expect("query schema before prune");
-        assert!(before.is_none());
+        assert_eq!(before.as_deref(), Some("subagentRun"));
 
         let result =
             subagent_history::prune_subagent_runs_for_parent_tool_calls(&conn, "conv-1", &[])
-                .expect("prune initializes subagent schema");
+                .expect("prune uses initialized subagent schema");
 
         assert_eq!(result.deleted_run_count, 0);
         let after: Option<String> = conn
