+++ crates/agent-gui/src-tauri/src/commands/mod.rs
@@ -3,6 +3,7 @@ pub mod cron;
 pub mod delegate;
 pub mod fs;
 pub mod gateway;
+pub mod history_db;
 pub mod hook;
 pub mod mcp;
 pub mod memory;
